GDPR-128 Adding config option to switch off data duplicate retention checks
Due to the data in certain environments, it does not make sense to
conduct data duplicate checks, data is annonymised in T3 for example.
This provides the option to switch those checks off where the data
or services are not available, allowing tests to still be carried
out on the rest of the process.