GDPR-113 Repeat pending deletion request (#154)
The Data Compliance service periodically requests a new batch
of offenders to be referred to it based on the offenders'
deletion dates being within a rolling window.

The concept of adding a limit to those returned has been
introduced, meaning that there could still be some offenders
left in that window that have not been referred yet. In this
case the request needs to be repeated using the same window
as before to ensure all eligible offenders are referred.